even stricter type system
Disallow FarPtr to carry pointer-to-pointer for too long.
It now tries to give the outermost pointer away, becoming
*FarPtr instead of FarPtr<*>.
Remove operator& that allowed to create such pointers: it
may be an illegal operation as far-pointer-to-far-pointer
may not even validly be created in a linear address spaces.
So the only way to create FarPtr<*> now is via FarPtrAsm::operator&,
which is guaranteed to be a valid operation (FarPtrAsm is
introduced exactly for that). But FarPtr created that way,
will try leak the pointer to the outside scope behind user's
back.