even stricter type system
Got rid of the leak-far-ptr technique by never allowing to
create FarPtr<*>, even from FarPtrAsm.
FarPtrAsm can now create AsmFarPtr with &, while the backward
conversion is possible with *. Also AsmFarPtr can create FarPtr<>*.

This pointed me to another bunch of wrong typecasts I added
previously, and actually forced me to revert them to the original
typecasts (or even lack of thereof!) of the freedos code.

So if the type systems starts to reject my previous changes
and forces to return the original code, then this is really
a good sign!