C++: remove typename=void hack
Got a lecture from C++ guru Anton Bikineev.
It appears the typename=void hack is not needed if you make
the template constraints a template parameter rather than the
default type for unnamed parameter. In this case the constraint
represents the so called "non-type template parameter", which
can be initialized with the constant value of that "non-type".
So add asterisk and use nullptr const as initializer
(could also write "enable_if<xxx, int> = 0" for example,
if asterisk hurts too much).

The difference is that the template parameters form the so-called
template signature, which must not clash for different templates.
std::enable_if<> template adds or removes the type, changing the
template signature, resulting in no clash.
OTOH the default value has no representation in the template
signature, so the presence or absence of the default value makes
no difference in the signature. As such, std::enable_if<> can't
help to avoid the clash if used as a default value.

OKey, never thought that meeting the world-famous C++ gurus is
needed to just write a basic templated stuff.