farptr: split main template
Create a minimal FarPtrBase template for __DOSFAR and
inherit the full-featured FarPtr from it that adds type
conversions.

The main point of this change is that only FarPtrBase
needs to be POD. FarPtr is no longer limited to being POD!
But in reality the PODness is still needed... Have to
investigate why.
I have to get rid of PODness to properly deal with dosobj stuff.