farptr: use shared_ptr to manage farobj
This is needed because even if I manually define the
assignment operator and delete obj before overwriting,
the problem is that the temporary object at the right
side of assignment will call the destructor, erasing
my object.
shared_ptr solves this and also allows to remove the
assignment operator, as it calls the destructor on an
object overwrite.