farptr: add _USE_FP() macro
This is a hack that stores the FP to the lut.
Needed for things like p->memb1.memb2.
In this case I don't see any simple way for 'memb2' to look up 'p'
(which is usually either in symtab or stored by MK_FP()),
so instead just store 'memb1' for now.