build: add pkgconfig support and move to lib/fdpp
It is not possible to put libfdpp.so to /usr/lib or lintian
will eat your ass. It will ask to to provide SONAME and 2 symlinks
(so.0.1, so.0, so), rename the package to libfdpp0-dev, and many other
things.
I've found it better to move to suffixed path, like /usr/lib/fdpp.
But this appears not that simple, as then autoconf loses the
ability to find your lib (and no, it doesn't have the handling
for custom pathes).
So I had to also add the pkgconfig support.