farptr: fix type corruption in ReadAllPartitionTables()
Problem found by ubsan.

This was because of wrong parsing of expression
*(char FAR * FAR *)MK_FP(0, 0x1e*4)

After dereference, we were getting a non-std layout object
in a dos space. After that, the conversion to char* was requested
by fmemcpy(), because it is actually a memcpy(). That last
conversion was happening via this invalid object, that should
normally have been __DOSFAR but was __FAR.
Added static asserts to avoid such situation in the future.
This static assert can't catch the bogus conversion itself,
as it was happening to char*, which is valid. But it was
preceeded with the conversion to FarPtr<char>* to satisfy
the dereference, as we don't provide the dereference op ourselves.
At this point we can catch the problem with assert.
It would be better to assert in constructors, but many of
them are default, so its not clear whether is possible.
But OTOH we can only get such wrong object by the static
cast, so assert at static cast makes sense.

Replaced the expression with the easier parsable equivalent.