config: fix MCB corruption with UMB at A0 [fixes #45]
The corruption was happening when conv mem was attempted to
enlarge to cover adjacent A0 UMB. prev_mcb() was not able to
go below uppermem_root, and was returning uppermem_root itself,
instead of its prev mcb. I added the assert to that: prev_mcb()
should not return the same value it was called with.
Make sure prev_mcb() is never called with uppermem_root as argument.