avoid calling hma_text before it is relocated [#78]
The problem was introduced by d3a7a357.
In principle we can call hma_text code before it is relocated.
But if it is overwritten by dyndata, then we suck.
So how can it be relocated if it is overwritten, you ask.
Well, it can. freedos makes the copy at start. So if the original
is overwritten, the copy can be perfectly used after relocation process.