config: fix tokenization [fixes #98]
Parser looks for digits and some special symbols at the
beginning of the line. But by mistake it was also treating
digits and special symbols the same way in the middle of
the line. Maybe ';' (comment sign) can be considered in the
middle of the line, but hope no one is ever doing that.
Certainly we don't want `shell=4dos.com` to become `shell=dos.com`.