rework user stack handling [fixes #97]
psp->ps_stack is now saved and preserved across the int21 calls.
return_user() is now done with the saved stack rather than with
parent's. Saved stack should be equal to parent's even when the
process self-parented his PSP, which crashes without that change.

This is yet another major departure from what FreeDOS does.
In fact this patch assumes that FreeDOS is not properly written.
If that assumption is wrong, then we'll get a regression. :)