chario: add kbd0a struct with larger buf [fixes #109]
int21/0a should allow up to 256 chars, but in freedos the
struct keyboard was defined to have only 129.
fdpp is paranoid on bounds checking, so it crashed.
Enlarging the main keyboard struct is unfortunately not an
option because it is defined in kernel.asm with 129-bytes buffer.
So introduce another one.

Adjust the code to allow its buffer to be near ptr.