rework exec_user() to fix multiple problems [fixes #105]
_int21_iret did Restore386Registers, for which exec_user()
doesn't have anything on stack. So it trashed the higher parts
of 32bit regs.
Introduce ret_user() for that and changed exec_user() to not
try to restore 32bit regs.

Some exe's (elite frontier) set up SP to some very small value,
so that the 24bytes iregs struct does not fit. In freedos this
caused the stack wrap-around and corruption of the top of the
64K segment where stack resides. But in fdpp the registers struct
is filled in protected mode so it wasn't splitted according to
the stack wrap-around. So it was also corrupting the next segment,
and was becoming partially unreachible for POPs.
Introduce smaller struct for passing only the essential regs
via the user stack. Hope there are no exe's that set SP to the
value smaller than 12...