task: change the termination of freed PSP [fixes #115]
It appears Alpha Waves game creates PSP with 0x26, then activates
it with 0x50, then resizes it to 0 with 0x4a (which presumably
means free?), then terminates it with 0x4c. And it does so many
times.
It expects that no resources are freed in this case.
And if PSP is freed, obviously trying to see what resources
it references, is unsafe.