procsupt: exec_user requires NORETURN [fixes #125]
exec_user() appeared to be special and can't work w/o NORETURN
attribute even for debugging purposes.
It can be called from init_DosExec() at boot, and init_DosExec()
uses the assembly stub (called INIT_DOSEXEC) that gets control
after exec completed. That stub calls exec_user() via int21/4bh.
If we remove NORETURN from exec_user(), then, when the program
terminates, the control is returned to the INIT_DOSEXEC stub.
It tries to return to init_DosExec(), but coopth instead awakes
exec_user() because they are running in the same coopthread.

So while, for debugging purposes, it is possible to remove NORETURN
from other functions, exec_user() will just crash, and there is
nothing to do to avoid that.