inthndlr: protect int21 from NULL derefs [fixes #151]
We could make these pointers nonnull instead, but unfortunately
currently nonnull is not always honoured. For example
c = *--fname; (dosfns.c:368)
would still crash because nonnull prop is lost after --.
We need to override ops in FarPtr template to recover nonnull
property, but its not frequently needed, so not worth the trouble.