dosfns: remove share checks when setting attrs
This was a very suspecious code that did share check
by trying to open file in RW COMPAT mode.
Maybe they meant to check for the cases where an RW
file would not be opened by 2 users, but RO will?
(such cases are marked with 1 and 2 in RBIL Table 01403).
So that clearing the RO attr would produce the sharing
violation...
But for such check to work, you need to check for exactly
2 existing openers, which seems to be not possible by
just trying to open the file in RW COMPAT mode...
Or maybe they wanted to make sure no one opens the file
in non-compat mode, before allowing to set attrs?
But that seems weird too, as I don't think set_attr should
fail only if the file is opened in non-compat mode, but
should succeed if the file is opened in compat mode...

So just remove that check completely.

Andrew, would you like to test if changing attrs is
possible while the file is open with various modes?