set up build for 186 instead of 386
The only difference is that the high words of 32bit
real-mode registers will not be saved/restored.
We do not need this as 32/64bit code of fdpp is all in protected
mode and therefore does not affect the real-mode registers.
OTOH the save/restore code is an ugly hack which shoud better
be disabled. In particular, it saves the high words below the
PSP-pointed stack pointer.

With that change, PSP:2eh will always point to the stack bottom.