share: implement proper critical error launcher [#152]
This fixes the testlock.exe crash.
Critical handler launcher should prepare the stack frame as
described in RBIL, which is what freedos share doesn't do.
So if the handler uses the return address to the caller of int21,
rather than the int24 return address, we get crash.

This patch creates the frame the way that both return addresses
returns from the critical handler but not from int21 itself, as
this cannot be handled by fdpp.

This means that the main CriticalError() (in entry.asm) should
also be adjusted to always return (so that the handler cannot
jump out of int21).