implement run-time data re-linking [closes #164]
This requires modifying symbols offsets, so add a new relocation
flags for offset modification.
The alternative could be to abuse PTR_MEMB macro, but that would
probably be quite convoluted as this macro can be created both
dynamically and statically, and we would need to distinguish these
cases.

Fixes https://github.com/dosemu2/dosemu2/issues/428