fatfs: fix cdsFlags check in get_dpb() [fixes #168]
Checking the CDSNETWDR flag was unreliable because it can be
set in a process of a file read, for example when config.sys is
being read and the redirector marks all drives as networked.
That resulted in a strange behaviour where the part of the file
that was cached can still be read, but eventually the cached
data finishes and you get EOF.

Check for CDSPHYSDRV flag instead.
Even if the drive is mapped via network, it can have the physical
drive (with fat) still being accessible.

An alternative (or a supplemental) fix can be to close (and later
re-open) config.sys after reading every line.