implement horrible relocation hack [#172]
As long as nasm lacks R_386_SEGRELATIVE support, we need to
stick to some hack that works.
This hack creates the relocation table by linking the kernel
to different addresses and comparing the resulting files.

This crazyness will be removed as soon as nasm has the needed
relocations support (aka NEVER).