remove binary kernel
We switched to ELF fully.
And binary kernel got already "broken" by those new elf-specific
sections, that would otherwise had to be filtered out with objcopy...

Just remove and forget.

Not incrementing the API version because dosemu2 already switched
to an elf API, so it won't be affected by the removal.