fix sfttbl termination condition [fixes #222]
According to this:
https://timobingmann.com/2006/helppc21/HelpPC-2.10-HTML/table-SFT.html

       00   dword      pointer to next system file table; offset of ffffh
                       indicates last table in chain

But freedos, and consequently fdpp, checked against the -1 pointer,
which is ffff:ffff. FreeDOS itself initializes the SFT table end as
ffff:ffff so that worked in most cases. Except when windows playes
with sft tables directly and changes them - then everything crashes
as it doesn't set the segment to 0xffff.