fix regression of e00daa95 [fixes #222]
Needs to reserve 43 bytes for dmatch struct, not 21, as in SDA.
It turns out sda_tmp_dm and alike structs overlap with the subsequent
data member SearchDir, so only 21 bytes reserved for dmatch.
But e00daa95 introduced a separate dmatch struct, and the size
was copy/pasted, leading to memory corruption.