init DOS_PSP with PGROUP [#172][skip CI]
We used to have DOS_PSP=cs, but now we have ELF headers in DOS space.
And as we use tiny model, cs:0 inevitably points to them.
To not overwrite them with PSP, move DOS_PSP var to kernel.asm and
initialize it with PGROUP.
This patch makes PGROUP no longer equal to other segments, so this
needs a proper segelf support. This can't be done in a master branch
where we relocate by a hard-coded value only (0x60).
So this patch shows a clear advantage of the segelf scheme usage.

Note that even the part of this patch, eg aligning PSP section to
0x100, cannot be back-ported to master because there we store PSP
at cs:0 regardless of the linker script, and overwrite whatever
is there. And there are the ELF headers now, since commit 95432f82.
Which probably shouldn't have been back-ported to master as a start.