entry: use IMR to try detect GPF [fixes #237]
Using ISR is unsafe because programs can legally chain to the prev
inthandler after EOI. We can rely on the fact that IRQ5 is masked
in IMR by default. If this is the case, then its most certainly a
GPF, so we can terminate the process. But if its not masked, always
pass to prev inthandler.