farptr: add safety wrapper for dot barriers
It may be needed if the barriers are added by some stupid script,
rather than manually as now.
Unfortunately, writing such script is going to be quite a challenge.
So far we only have very few places where this fixup was needed,
so automating the task may not worth the trouble, but lets add the
wrapper nevertheless, then see.