lfn: Implement int21/71a0 get volume information
If the volume is local then return suitable information, as either
FAT or FAT32. Currently FDPP and FreeDOS don't support LFN
natively, so values are chosen to reflect that. If DOSLFN is loaded
then its own handler will used in preference.

If the disk is remote then query the respective redirector via
int2f/11a0 (an extension) for the relevant values. If the redirector
dosen't support the extension, then we pretend that the kernel doesn't
support int21/71a0.