emit unwrapped aggregates explicitly [#259]
Had to add manual fixups around the freedos code.
But this is a relatively small price for emitting the refs w/o
the C++ ability to overload operator dot.

The bigger problem is a placement-new that this patch adds.
It will not be removed by subsequent patches, even though all
other placement-new's - will...