chario: fix fdpp abort in read_line_handle()
read_line_handle() was calling do_read_line(), doing the inline
cast of far ptr arg. This inline cast was holding a reference to
an object, which resulted in a fact that all writes to that object
from do_read_line() were treated by the write checker as a single
expression writes!

Move the cast out of the function call.
This way it no longer holds the reference while call is in progress.